+++ src/cli/build.helpers.ts
@@ -11,51 +11,18 @@ import { promises as fs } from "fs";
 import { generateAllYaml } from "../lib/assets/yaml/generateAllYaml";
 import { webhookConfigGenerator } from "../lib/assets/webhooks";
 import { generateZarfYamlGeneric } from "../lib/assets/yaml/generateZarfYaml";
-import { ModuleConfig } from "../lib/core/module";
 
-export type PeprNestedFields = Pick<
-  ModuleConfig,
-  | "uuid"
-  | "onError"
-  | "webhookTimeout"
-  | "customLabels"
-  | "alwaysIgnore"
-  | "env"
-  | "rbac"
-  | "rbacMode"
-> & {
-  peprVersion: string;
-};
-
-export type LoadModuleReturn = {
-  cfg: PeprConfig;
-  entryPointPath: string;
-  modulePath: string;
-  name: string;
-  path: string;
-  uuid: string;
-};
-
-export type PeprConfig = Omit<ModuleConfig, keyof PeprNestedFields> & {
-  pepr: PeprNestedFields & {
-    includedFiles: string[];
-  };
-  description: string;
-  version: string;
-};
-
-export type BuildModuleReturn = {
-  ctx: BuildContext<BuildOptions>;
-  path: string;
-  cfg: PeprConfig;
-  uuid: string;
-};
 interface ImageOptions {
   customImage?: string;
   registryInfo?: string;
   peprVersion?: string;
   registry?: string;
 }
+/**
+ * Assign image string
+ * @param imageOptions CLI options for image
+ * @returns image string
+ */
 export function assignImage(imageOptions: ImageOptions): string {
   const { customImage, registryInfo, peprVersion, registry } = imageOptions;
   if (customImage) {
@@ -71,12 +38,6 @@ export function assignImage(imageOptions: ImageOptions): string {
   }
 
   return "";
-}
-export function shouldExitEarly(buildModuleResult: BuildModuleReturn): boolean {
-  if (buildModuleResult?.cfg && buildModuleResult.path && buildModuleResult.uuid) {
-    return false;
-  }
-  return true;
 }
 
 export type Reloader = (opts: BuildResult<BuildOptions>) => void | Promise<void>;
